// src/types/types.ts
export type SkillKey =
  | 'AirMagic'
  | 'Archery'
  | 'Armorer'
  | 'Artillery'
  | 'Ballistics'
  | 'Diplomacy'
  | 'EagleEye'
  | 'EarthMagic'
  | 'Estates'
  | 'FireMagic'
  | 'FirstAid'
  | 'Intelligence'
  | 'Leadership'
  | 'Learning'
  | 'Logistics'
  | 'Luck'
  | 'Mysticism'
  | 'Navigation'
  | 'Necromancy'
  | 'Offense'
  | 'Pathfinding'
  | 'Resistance'
  | 'Scholar'
  | 'Scouting'
  | 'Sorcery'
  | 'Tactics'
  | 'WaterMagic'
  | 'Wisdom'
  | 'Interference'